extract styles from dashboard into it's own file
