+++ src/components/dashboard/dashboard.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { View, Image, TextInput, StyleSheet, TouchableOpacity, Text, AsyncStorage, StatusBar, Keyboard} from 'react-native';
+import styles from '../../../src/styles/dashboard/dashboardStyles'
 
 export default class Dashboard extends Component {
 
@@ -90,43 +91,3 @@ export default class Dashboard extends Component {
     )
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#075473'
-  },
-  logoContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginTop: 30
-  },
-  logo: {
-    width: 120,
-    height: 120
-  },
-  dashboardRow: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    paddingHorizontal: 60,
-    paddingVertical: 60
-  },
-  dashboardIcon: {
-    width: 125,
-    height: 125,
-  },
-  header: {
-    color: '#FFF',
-    marginTop: 5,
-    textAlign: 'center',
-    opacity: 0.7
-  },
-  info: {
-    color: '#FFF',
-    marginTop: 10,
-    textAlign: 'center',
-    opacity: 0.7,
-    fontSize: 15
-  }
-})
